[C8] Split serialize_restore.go per entity family (#2244)
model/sketch/serialize_restore.go was 611 LOC, over the 500-line ceiling.
Split the sketch recipe restore per entity family; each file is now under the
ceiling:

- serialize_restore.go (336): the restore entry (ApplyRecipe/restoreSketch),
  the sketchRestorer state, and the shared reference re-binding helpers
  (points/point/entity/line/arc/curve/smooth/at, ellipse/offsetSpline, the
  two-points/two-lines/two-curves/point-and-line binders, pending drives).
- serialize_restore_entity.go (148): the point/entity restore (points,
  entities, projected point/curve, spline extras, plane).
- serialize_restore_constraint.go (155): the constraint/dimension restore
  (geometric constraints, basic + advanced dimensions).

Behaviour-preserving: split programmatically by receiver-qualified
declaration, verified every non-comment code line is byte-identical (only
per-file import redistribution differs); imports resolved per file with
goimports. serialize_*.go is already funlen/gocyclo-excluded by path. Sketch
package builds + tests pass, archguard passes, lint 0. Net delta of
recognizers / tolerance constants / fallback sites / type assertions: 0.

Part of #2183 (C8: structure/size limits). Closes #2244.

(cherry picked from commit 98bf39de74b0fb168bc1ba7b22c875798917a860)
