+++ model/sketch/serialize_restore.go
@@ -5,7 +5,6 @@ package sketch
 import (
 	"fmt"
 
-	"oblikovati.org/api/types"
 	"oblikovati.org/math"
 	"oblikovati.org/model/linetype"
 	"oblikovati.org/model/seq"
@@ -139,272 +138,11 @@ func (r *sketchRestorer) pin(obj idCarrier, saved int) {
 	}
 }
 
-func (r *sketchRestorer) restorePoints(points []PointData) {
-	for _, pd := range points {
-		r.pointMap[pd.ID] = r.restorePoint(pd)
-	}
-}
-
-// restorePoint recreates one point (standalone SketchPoint or a curve-owned point) and
-// pins its persisted local id.
-func (r *sketchRestorer) restorePoint(pd PointData) *Point {
-	pos := math.P2(pd.X, pd.Y)
-	p := r.newPointFor(pd.Standalone, pos)
-	p.SetCenterPoint(pd.CenterPoint) // hole-centre marker (#2015)
-	r.pin(p, pd.ID)
-	return p
-}
-
-// newPointFor creates a standalone SketchPoint or a curve-owned solver point, exactly one.
-func (r *sketchRestorer) newPointFor(standalone bool, pos math.Point2) *Point {
-	if standalone {
-		return r.s.points.Add(pos)
-	}
-	return r.s.newPoint(pos)
-}
-
-func (r *sketchRestorer) restoreEntities(entities []EntityData) error {
-	for _, ed := range entities {
-		e, err := r.restoreEntity(ed)
-		if err != nil {
-			return err
-		}
-		// Projected reference entities (#1268) carry no construction flag and pin their own ids,
-		// so each post-step is optional rather than a hard type-assert.
-		if sc, ok := e.(interface{ SetConstruction(bool) }); ok {
-			sc.SetConstruction(ed.Construction)
-		}
-		if ed.Centerline {
-			if cl, ok := e.(interface{ SetCenterline(bool) }); ok {
-				cl.SetCenterline(true)
-			}
-		}
-		if ic, ok := e.(idCarrier); ok {
-			r.pin(ic, ed.ID)
-		}
-		r.s.readEntityFormat(ed, e.EntityID()) // per-entity format overrides (#2015)
-		r.entityMap[ed.ID] = e
-	}
-	return nil
-}
-
-// restoreEntity rebuilds one entity through its kind's registered codec — the
-// same pairing its serializeEntity encode came from, so the two can never
-// drift (#1624). An unknown kind is a hard error: a recipe never restores
-// partially.
-func (r *sketchRestorer) restoreEntity(ed EntityData) (Entity, error) {
-	c, ok := entityCodecs2D[EntityKind(ed.Kind)]
-	if !ok {
-		return nil, fmt.Errorf("unknown entity kind %q", ed.Kind)
-	}
-	return c.decode(r, ed)
-}
-
-// restoreProjectedPoint rebuilds a frozen projected point and registers its anchor in the point
-// map so constraints referencing the anchor restore, pinning the anchor id (#1268).
-func (r *sketchRestorer) restoreProjectedPoint(ed EntityData) (Entity, error) {
-	if len(ed.Points) < 1 || len(ed.Anchor) < 2 {
-		return nil, fmt.Errorf("projectedPoint needs an anchor id and a 2-component anchor")
-	}
-	pos := math.P2(math.Scalar(ed.Anchor[0]), math.Scalar(ed.Anchor[1]))
-	pp := r.s.RestoreProjectedPoint(ID(ed.Points[0]), pos, ed.SourceKind, ed.Source)
-	r.pointMap[ed.Points[0]] = pp.Anchor()
-	r.note(ed.Points[0])
-	r.note(ed.ID)
-	return pp, nil
-}
-
-// restoreProjectedCurve rebuilds a projected curve as its concrete grounded reference entity
-// (ADR-0055 phase 3): it re-creates the reference Line/Circle/Arc/Ellipse/Spline from the persisted
-// analytic form (or polyline), re-registers the frozen Projection that drives it, and pins the
-// entity's defining points to their saved ids in the point map so constraints referencing them
-// restore. The entity's own id is pinned by restoreEntities' generic post-step.
-func (r *sketchRestorer) restoreProjectedCurve(ed EntityData) (Entity, error) {
-	ent, ok := r.s.buildReferenceEntity(ed.ProjShape, ed.ProjParams, ed.Coords)
-	if !ok {
-		return nil, fmt.Errorf("projectedCurve %d: cannot rebuild reference entity (shape %q, %d params, %d coords)",
-			ed.ID, ed.ProjShape, len(ed.ProjParams), len(ed.Coords))
-	}
-	r.s.RestoreProjection(ent, ed.SourceKind, ed.Source)
-	dps := DefiningPoints(ent)
-	if len(ed.Points) != len(dps) {
-		return nil, fmt.Errorf("projectedCurve %d: %d point ids for a %d-point reference entity", ed.ID, len(ed.Points), len(dps))
-	}
-	for i, p := range dps {
-		p.setID(ID(ed.Points[i]))
-		r.pointMap[ed.Points[i]] = p
-		r.note(ed.Points[i])
-	}
-	r.note(ed.ID)
-	return ent, nil
-}
-
 // note raises the restorer's max-id watermark for an id pinned outside pin() (the
 // self-id'd projected entities), so later minted ids never collide with restored ones.
 func (r *sketchRestorer) note(id int) {
 	if uint64(id) > r.maxID {
 		r.maxID = uint64(id)
-	}
-}
-
-func (r *sketchRestorer) restoreConstraints(constraints []ConstraintData) error {
-	for _, cd := range constraints {
-		if err := r.restoreConstraint(cd); err != nil {
-			return fmt.Errorf("constraint %q: %w", cd.Kind, err)
-		}
-	}
-	return nil
-}
-
-// restoreConstraint decodes through the paired codec registry (#1625) — an
-// unknown kind is a corrupt-recipe error, named honestly.
-func (r *sketchRestorer) restoreConstraint(cd ConstraintData) error {
-	codec, ok := constraintCodecs2D[ConstraintKind(cd.Kind)]
-	if !ok {
-		return fmt.Errorf("unknown constraint kind %q", cd.Kind)
-	}
-	return codec.decode(r, cd)
-}
-
-// restoreSplineExtras rebuilds a spline's fit method and active tangency
-// handles (M06-F11, #626).
-func restoreSplineExtras(s *Sketch, sp *Spline, ed EntityData) error {
-	if ed.FitMethod != "" {
-		m, ok := types.ParseSplineFitMethod(ed.FitMethod)
-		if !ok {
-			return fmt.Errorf("unknown spline fit method %q (want smooth|sweet|chord)", ed.FitMethod)
-		}
-		sp.FitMethod = m
-	}
-	for _, hd := range ed.Handles {
-		h, err := s.splineHandles.Activate(sp, hd.FitIndex)
-		if err != nil {
-			return err
-		}
-		h.End.SetPosition(math.P2(math.Scalar(hd.EndX), math.Scalar(hd.EndY)))
-	}
-	return nil
-}
-
-func (r *sketchRestorer) restoreDimensions(dims []DimensionData) error {
-	for _, dd := range dims {
-		d, err := r.restoreDimension(dd)
-		if err != nil {
-			return fmt.Errorf("dimension %q: %w", dd.Kind, err)
-		}
-		if dd.Driven {
-			d.SetDriven(true)
-		}
-		if dd.Limits != nil {
-			d.SetLimits(dd.Limits.Min, dd.Limits.Max)
-		}
-		if len(dd.TextPoint) == 2 {
-			d.SetTextPoint(math.P2(math.Scalar(dd.TextPoint[0]), math.Scalar(dd.TextPoint[1])))
-		}
-	}
-	return nil
-}
-
-func (r *sketchRestorer) restoreDimension(dd DimensionData) (*DimensionConstraint, error) {
-	dc := r.s.dimCons
-	switch dd.Kind {
-	case "distance":
-		a, err := r.point(dd.Points, 0)
-		if err != nil {
-			return nil, err
-		}
-		b, err := r.point(dd.Points, 1)
-		if err != nil {
-			return nil, err
-		}
-		o, ok := ParseDistanceOrientation(dd.Orientation)
-		if !ok {
-			return nil, fmt.Errorf("distance dimension: unknown orientation %q", dd.Orientation)
-		}
-		return dc.AddDistanceOriented(a, b, dd.Expression, o)
-	case "radius":
-		// AddRadius accepts any CircularCurve (circle or arc) — an arc radius/diameter
-		// dimension is common (esp. from CAD imports), so resolve as a circular curve,
-		// not a *Circle. Restoring via r.circle rejected radius-dimensioned arcs and
-		// blocked ~40% of real Inventor-exported parts. Symmetric with the write side,
-		// which serializes d.refs regardless of circle-vs-arc.
-		c, err := r.curve(dd.Curves, 0)
-		if err != nil {
-			return nil, err
-		}
-		return dc.AddRadius(c, dd.Expression)
-	case "diameter":
-		c, err := r.curve(dd.Curves, 0)
-		if err != nil {
-			return nil, err
-		}
-		return dc.AddDiameter(c, dd.Expression)
-	case "angle":
-		l1, err := r.line(dd.Curves, 0)
-		if err != nil {
-			return nil, err
-		}
-		l2, err := r.line(dd.Curves, 1)
-		if err != nil {
-			return nil, err
-		}
-		return dc.AddAngle(l1, l2, dd.Expression)
-	case "arcLength":
-		a, err := r.arc(dd.Curves, 0)
-		if err != nil {
-			return nil, err
-		}
-		return dc.AddArcLength(a, dd.Expression)
-	default:
-		return r.restoreAdvancedDimension(dd)
-	}
-}
-
-// restoreAdvancedDimension rebuilds the M21 dimension kinds (offset/three-point-angle/
-// ellipse-radius); split out of restoreDimension to keep that switch small.
-func (r *sketchRestorer) restoreAdvancedDimension(dd DimensionData) (*DimensionConstraint, error) {
-	dc := r.s.dimCons
-	switch dd.Kind {
-	case "offsetDim":
-		p, err := r.point(dd.Points, 0)
-		if err != nil {
-			return nil, err
-		}
-		l, err := r.line(dd.Curves, 0)
-		if err != nil {
-			return nil, err
-		}
-		return dc.AddOffsetDim(p, l, dd.LinearDiameter, dd.Expression)
-	case "threePointAngle":
-		pts, err := r.points(dd.Points, 3)
-		if err != nil {
-			return nil, err
-		}
-		return dc.AddThreePointAngle(pts[0], pts[1], pts[2], dd.Expression)
-	case "ellipseRadius":
-		e, err := r.ellipse(dd.Curves, 0)
-		if err != nil {
-			return nil, err
-		}
-		return dc.AddEllipseRadius(e, dd.Expression)
-	case "tangentDistance":
-		l, err := r.line(dd.Curves, 0)
-		if err != nil {
-			return nil, err
-		}
-		c, err := r.curve(dd.Curves, 1)
-		if err != nil {
-			return nil, err
-		}
-		return dc.AddTangentDistance(l, c, dd.FarSide, dd.LinearDiameter, dd.Expression)
-	case "offsetSplineDim":
-		o, err := r.offsetSpline(dd.Curves, 0)
-		if err != nil {
-			return nil, err
-		}
-		return dc.AddOffsetSplineDim(o, dd.Expression)
-	default:
-		return nil, fmt.Errorf("unknown dimension kind %q", dd.Kind)
 	}
 }
 
@@ -595,17 +333,4 @@ func at(ids []int, i int, what string) (int, error) {
 		return 0, fmt.Errorf("missing %s operand %d (have %d)", what, i, len(ids))
 	}
 	return ids[i], nil
-}
-
-// restorePlane rebuilds a sketch plane from its serialized origin and axes.
-func restorePlane(pd PlaneData) (Plane, error) {
-	x, err := math.UnitVector3FromVector(vector3(pd.XAxis))
-	if err != nil {
-		return Plane{}, fmt.Errorf("plane x-axis: %w", err)
-	}
-	y, err := math.UnitVector3FromVector(vector3(pd.YAxis))
-	if err != nil {
-		return Plane{}, fmt.Errorf("plane y-axis: %w", err)
-	}
-	return NewPlane(math.P3(pd.Origin[0], pd.Origin[1], pd.Origin[2]), x, y)
 }
